refactor: split source_graph.py to fix CI file-size hard-limit breach
Merging origin/main (which independently grew source_graph.py to 1993
lines via unrelated ADR-041 P0 work) with this branch's small
_common_prefix_len addition pushed the file to 2077 lines, tripping
the AI-readiness gate's 2000-line hard cap and failing unit-tests on
every CI lane.

Split Phase 5's finding-emission logic (_diff_source_graph_findings
and its per-family helpers) out to a new source_graph_findings.py,
mirroring the inline.py/inline_graph_fold.py split already established
in this package. source_graph.py keeps the graph schema/construction
(GraphNode/GraphEdge/SourceGraphSummary, build_source_graph, the
structural diff_source_graph) and re-exports diff_source_graph_findings
via a lazy __getattr__ shim (same pattern as cli_buildsource.py's
existing back-compat shim) so existing callers importing it from
source_graph keep working without a static import cycle.
